+++ x/cronos/genesis_test.go
@@ -36,6 +36,78 @@ func (suite *CronosTestSuite) TestInitGenesis() {
 				}},
 			true,
 		},
+		{
+			"Wrong denom in external token mapping",
+			func() {},
+			&types.GenesisState{
+				ExternalContracts: []types.TokenMapping{
+					{
+						Denom:    "aaa/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865",
+						Contract: "0x0000000000000000000000000000000000000000",
+					},
+				},
+			},
+			true,
+		},
+		{
+			"Wrong denom in auto token mapping",
+			func() {},
+			&types.GenesisState{
+				AutoContracts: []types.TokenMapping{
+					{
+						Denom:    "aaa/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865",
+						Contract: "0x0000000000000000000000000000000000000000",
+					},
+				},
+			},
+			true,
+		},
+		{
+			"Wrong contract in external token mapping",
+			func() {},
+			&types.GenesisState{
+				ExternalContracts: []types.TokenMapping{
+					{
+						Denom:    "ibc/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865",
+						Contract: "0x00000000000000000000000000000000000000",
+					},
+				},
+			},
+			true,
+		},
+		{
+			"Wrong contract in auto token mapping",
+			func() {},
+			&types.GenesisState{
+				AutoContracts: []types.TokenMapping{
+					{
+						Denom:    "ibc/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865",
+						Contract: "0x00000000000000000000000000000000000000",
+					},
+				},
+			},
+			true,
+		},
+		{
+			"Correct token mapping",
+			func() {},
+			&types.GenesisState{
+				Params: types.DefaultParams(),
+				ExternalContracts: []types.TokenMapping{
+					{
+						Denom:    "ibc/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865",
+						Contract: "0x0000000000000000000000000000000000000000",
+					},
+				},
+				AutoContracts: []types.TokenMapping{
+					{
+						Denom:    "gravity0x0000000000000000000000000000000000000000",
+						Contract: "0x0000000000000000000000000000000000000000",
+					},
+				},
+			},
+			false,
+		},
 	}
 
 	for _, tc := range testCases {